backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Stanyslas250 <stanyslas.contact@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
sqlalchemy = "^2.0.31"
sqlmodel = "^0.0.21"
pydantic-settings = "^2.3.4"
passlib = "^1.7.4"
psycopg2 = "^2.9.9"
alembic = "^1.13.2"
pyjwt = "^2.8.0"
tenacity = "^8.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
